requirements.txt
appdirs==1.4.3
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
importlib-metadata==1.5.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.1
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
py==1.8.1
pybind11==2.4.3
pycuda==2019.1.2
pygame==1.9.6
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
pytools==2020.1
six==1.14.0
taichi-nightly==0.5.2
tqdm==4.43.0
typed-ast==1.4.1
wcwidth==0.1.8
wrapt==1.11.2
zipp==3.0.0
